Initial commit
Revise Lab3B code to replace labrpc with the actual, networked
counterpart. Keep labgob, but delete other 6.824 machinery designed
for local testing in favor of distributed shims + a configuration
interface, to be written.

Note that this code does not yet run. There is no main, there is
no Makefile. There is no driver that will accurately allow us to
test kv in vitro.

Do not yet modify the persister: we will test this down the road.
Networking is more important.

Signed-off-by: Jay Lang <jaytlang@mit.edu>