First pass at networked KV driver
Add a main function + build utilities. Currently, the binary should dial
a fixed list of peers (hardcoded into the netdrv package), and then spin
up a KV server as seen during the labs. Clients will have the
opportunity to interact with the clerk via CLI.

Nits: Fix the README and polish the temporary details of netconfig.

Signed-off-by: Jay Lang <jaytlang@mit.edu>