Untested: add in locking opcodes
In the client, add in acquire and release primitives, which gain
and release the lock name passed, respectively. Acquire will
spin until the lock is acquired; release will always return
successfully regardless of whether we hold the lock or not.

In the server, add four new opcodes. AcquireOp and ReleaseOp will
attempt to acquire in commitOp and release in commitOp, respectively.
Make commitOp return an error, such that if this error is detected
(i.e. an acquire or release fails) the opcode of the command is
CHANGED to reflect this before broadcasting. Thus, applyCh watchers
pick this up and pass the correct error message back to the client.

We attempt to correct the situation where an acquire is committed but
not yet broadcasted to clients by checking the current value of the
lock. This intuitively makes sense, but I'm not sure if it's totally
correct. Testing to come.

Signed-off-by: Jay Lang <jaytlang@mit.edu>