Add leasing, with a guess of 5 seconds
Make it so if a lock is held for 5 seconds by a client, the client
is assumed down and the lock is released. This is crude, but
nobody should hold a lock for more than 5 seconds.

If the above assumption turns out bad, we can set up a more complex
data structure to renew leases should a client really need a lock
for more time. This is unlikely, since we impose the invariant
inside client code that a kernel thread can only hold one block
at a time, and blocks are small.

Signed-off-by: Jay Lang <jaytlang@mit.edu>