First pass at journaling
Write a ext3-style log, which allows us to make atomic writes
of several blocks in size at once. This code has yet to be tested,
but should provide atomicity in the face of crashes at any point.
Initialization code has been modified to replay the log, which is
trimmed whenever no outstanding system calls are using the log's
superblock.

Signed-off-by: Jay Lang <jaytlang@mit.edu>