Add handling for indirect blocks, in preparation for a fuller inode API
Write indirect blocks, which are a list of uints, in. The indirect API
checks to make sure that not too many indirects happen at once, and
returns a list of uints which can in turn be grabbed via the block
API.

Additionally, add semantics so that inode users have a uniform spec
of what may fail and when. Again, if a failure occurs, the entire
transaction must be rolled back.

Signed-off-by: Jay Lang <jaytlang@mit.edu>