Change balloc to accept multiple blocks at a time
Since each invocation of balloc() calls getBitmap() internally,
we need to ensure it only gets called once at a time, since
during each transaction we will get the view of the bitmap
as it occurred before the transaction. To duct tape this together
with calls like AllocNextBlock(), which requires that we might
alloc both an indirect and the first block of said indirect, we
need an adjustable parameter here.

Signed-off-by: Jay Lang <jaytlang@mit.edu>