Eliminate the indirect block
The indirect causes problems with our delayed write passage.  In
particular, the problematic situation is in the inode block allocation
routine, and it's theoretically resolvable but clunky. When you start
below NDIRECT and go over NDIRECT, you need to allocate a new indirect
block and specify its block number, and you also need to populate that
indirect block with new block numbers. Doing those two things together
is feasible, but then the caller can't read the indirect since it
contains stale data (even if the indirect was flushed to the log). This
means you have to return an arbitrarily large list of blocks, all of
which become locked, which the client needs to manage.

Accordingly, eliminate the indirect and instead use a larger number
of direct blocks (abusing our block size + 1 inode/block thing) to
achieve a filesystem size of 64G and max filesize of 1022 blocks,
or about 4 megs.

Signed-off-by: Jay Lang <jaytlang@mit.edu>