Start testing
Write and verify some of the basic tests. At this rate, I've
realized that testing the retry semantics of acquire will
be difficult, and the mock disk will have to be multithreaded
(e.g. so I can simulate two clients attacking it and verify
that one client gets the lock after another).

Signed-off-by: Jay Lang <jaytlang@mit.edu>